% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_invalidation_tables.R
\name{import_invalidation_tables}
\alias{import_invalidation_tables}
\title{Function to import invalidation data from the \code{`invalidations`} and
\code{`invalidations_date_ranges`} tables as a single table for a ssensors 
\strong{smonitor} database.}
\usage{
import_invalidation_tables(con, tz = "UTC")
}
\arguments{
\item{con}{Database connection to a ssensors \strong{smonitor} database.}

\item{tz}{Time zone for the dates to be parsed into. Default is \code{"UTC"}.}
}
\value{
Tibble.
}
\description{
Function to import invalidation data from the \code{`invalidations`} and
\code{`invalidations_date_ranges`} tables as a single table for a ssensors 
\strong{smonitor} database.
}
\seealso{
\code{\link{import_invalidations}}, 
\code{\link{import_invalid_date_ranges}}
}
\author{
Stuart K. Grange
}
