% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_meteorological_data_by_site.R
\name{import_meteorological_data_by_site}
\alias{import_meteorological_data_by_site}
\title{Function to import meteorological data for a monitoring site from an 
extended sensors \strong{smonitor} database.}
\usage{
import_meteorological_data_by_site(
  con,
  site,
  summary = 1L,
  start = 1969,
  end = NA,
  variables = c("air_temp", "rh", "wd", "ws", "global_radiation"),
  valid_only = FALSE,
  set_invalid_values = FALSE,
  tz = "UTC",
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{con}{Database connection to an sensors \strong{smonitor} database.}

\item{site}{A vector of sites. If the \code{`sites`} table has a 
\code{site_pair} variable, this will be used as a switch.}

\item{summary}{A vector of summaries.}

\item{start}{What is the start date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and
will floor-rounded.}

\item{end}{What is the end date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and 
will be ceiling-rounded.}

\item{variables}{What meteorological variables should be queried?}

\item{valid_only}{Should invalid observations be filtered out?}

\item{set_invalid_values}{Should invalid observations be set to \code{NA}?}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Tibble.
}
\description{
Function to import meteorological data for a monitoring site from an 
extended sensors \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
