% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites_met.R
\name{import_sites_met}
\alias{import_sites_met}
\title{Function to import meteorological data for a monitoring site from an 
extended sensors \strong{smonitor} database.}
\usage{
import_sites_met(
  con,
  site,
  summary = 1L,
  start = 1969,
  end = NA,
  variables = c("air_temp", "rh", "wd", "ws"),
  valid_only = FALSE,
  tz = "UTC"
)
}
\arguments{
\item{con}{Database connection to an sensors \strong{smonitor} database.}

\item{site}{A vector of sites. If the \code{`sites`} table has a 
\code{site_pair} variable, this will be used as a switch.}

\item{summary}{A vector of summaries.}

\item{start}{What is the start date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and
will floor-rounded.}

\item{end}{What is the end date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and 
will be ceiling-rounded.}

\item{variables}{What meteorological variables should be queried?}

\item{valid_only}{Should invalid observations be filtered out?}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}
}
\value{
Tibble.
}
\description{
Function to import meteorological data for a monitoring site from an 
extended sensors \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
