% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_with_sensor_id.R
\name{import_with_sensor_id}
\alias{import_with_sensor_id}
\alias{import_reference_cylinders}
\alias{import_cylinder_deployments}
\alias{import_observations_calibrations}
\alias{import_cylinder_test_summaries}
\alias{import_sensor_deployments}
\alias{import_calibration_coefficients}
\alias{import_sensing_elements}
\alias{import_observation_flags}
\alias{import_cylinder_exclusion_periods}
\alias{import_sensors}
\alias{import_sensor_types}
\alias{import_invalid_date_ranges}
\alias{import_validity_types}
\alias{import_model_objects}
\alias{import_raster_objects}
\alias{import_observation_flagging_conditions}
\alias{import_generic_observations}
\title{Functions to import data from an extended sensors \strong{smonitor} database.}
\usage{
import_with_sensor_id(
  con,
  process,
  summary = NA,
  start = 1969,
  end = NA,
  site_name = TRUE,
  valid_only = FALSE,
  warn = FALSE,
  tz = "UTC"
)

import_reference_cylinders(con, tz = "UTC")

import_cylinder_deployments(con, tz = "UTC")

import_observations_calibrations(con, tz = "UTC")

import_cylinder_test_summaries(con, tz = "UTC")

import_sensor_deployments(con, tz = "UTC")

import_calibration_coefficients(con, tz = "UTC")

import_sensing_elements(con, tz = "UTC")

import_observation_flags(con)

import_cylinder_exclusion_periods(con, tz = "UTC")

import_sensors(con)

import_sensor_types(con)

import_invalid_date_ranges(con, tz = "UTC")

import_validity_types(con)

import_model_objects(con, add_extras = TRUE)

import_raster_objects(con, tz = "UTC")

import_observation_flagging_conditions(con)

import_generic_observations(con, tz = "UTC")
}
\arguments{
\item{con}{Database connection to an sensors \strong{smonitor} database.}

\item{process}{A vector of processes.}

\item{summary}{A vector of summaries.}

\item{start}{What is the start date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and
will floor-rounded.}

\item{end}{What is the end date of data to be returned? Ideally, the 
date format should be \code{yyyy-mm-dd}, but the UK locale convention of 
\code{dd/mm/yyyy} will also work. Years as strings or integers work too and 
will be ceiling-rounded.}

\item{site_name}{Should the return include the \code{site_name} variable? 
Default is \code{TRUE}.}

\item{valid_only}{Should invalid observations be filtered out?}

\item{warn}{Should the functions raise warnings?}

\item{tz}{Time-zone for the dates to be parsed into. Default is \code{"UTC"}.}

\item{add_extras}{Should extra variables be calculated and returned?}
}
\value{
Tibble.
}
\description{
Functions to import data from an extended sensors \strong{smonitor} database.
}
\seealso{
\link{import_by_process}
}
\author{
Stuart K. Grange
}
