% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sensor_ids.R
\name{update_sensor_ids}
\alias{update_sensor_ids}
\title{Function to update the \code{`sites`}'s \code{sensor_ids} variable for a 
sensors \strong{smonitor} database.}
\usage{
update_sensor_ids(con)
}
\arguments{
\item{con}{Database connection to an sensors \strong{smonitor} database.}
}
\value{
Invisible \code{con}.
}
\description{
Function to update the \code{`sites`}'s \code{sensor_ids} variable for a 
sensors \strong{smonitor} database.
}
\author{
Stuart K. Grange
}
